+++ release-notes.md
@@ -23,6 +23,7 @@ For modders:
 * You can now specify minimum dependency versions in `manifest.json`.
 * Added `System.ValueTuple.dll` to the SMAPI install package so mods can use [C# 7 value tuples](https://docs.microsoft.com/en-us/dotnet/csharp/tuples).
 * Improved trace logging when SMAPI loads mods.
+* Changed `manifest.MinimumApiVersion` from string to `ISemanticVersion`.
 * Fixed `SemanticVersion` parsing some invalid versions into close approximations (like `1.apple` &rarr; `1.0-apple`).
 * Fixed `SemanticVersion` not treating hyphens as separators when comparing prerelease tags.  
   <small>_(While that was technically correct, it leads to unintuitive behaviour like sorting `-alpha-2` _after_ `-alpha-10`, even though `-alpha.2` sorts before `-alpha.10`.)_</small>